refactor(renderer/pages/game-details): colocated repacks files to one feature folder
