+++ src/renderer/src/pages/game-details/repacks/index.ts
@@ -1,0 +1,1 @@
+export * from "./repacks-modal";
