+++ src/renderer/src/pages/game-details/repacks/repacks-modal.tsx
@@ -6,9 +6,9 @@ import type { GameRepack } from "@types";
 
 import * as styles from "./repacks-modal.css";
 
-import { SPACING_UNIT } from "../../theme.css";
+import { SPACING_UNIT } from "../../../theme.css";
 import { format } from "date-fns";
-import { SelectFolderModal } from "./select-folder-modal";
+import { SelectFolderModal } from "../select-folder-modal";
 
 export interface RepacksModalProps {
   visible: boolean;
